zrusenie paralelneho spracovania grafov pre jednotve vetvy

+++ VypocetKostier.h
@@ -1,3 +1,4 @@
+
 #pragma once
 #include <stdio.h>
 #include <vector>
@@ -9,11 +10,11 @@ class VypocetKostier
 {
     static double maxH(const std::vector<std::vector<double>>& hodnoty);
     using Matica = std::vector<std::vector<double>>;
-    static long gauss(Matica& matica, int n);
-    static long kofaktor(Matica& matica, int n, int r, int s);
+    static long long gauss(Matica& matica, int n);
+    static unsigned long long kofaktor(Matica& matica, int n, int r, int s);
     
 public:
     
     using Hrany = std::vector<std::vector<int>>;
-    static long celkovyVypocet(const Hrany& hrany, int reg, int n);
+    static unsigned long long celkovyVypocet(const Hrany& hrany, int reg, int n);
 };
